Point the target to Trace Compass 0.1.0
The Trace Compass "stable" repository is continuously updated to point
to the latest stable release. Right now it contains the 1.1.0 one.

TotalADS does not compile against this anymore, so instead point to
the specific 0.1.0 repository (which is the latest one that works)
to avoid this problem.

Once the code is updated, the target could be changed again to point
to 1.1.0, and so on. Pointing to "stable", which can change anytime,
is dangerous!

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>